refact: move Status out of its own package
