+++ core/mastodon/src/main/java/com/jeanbarrossilva/orca/core/mastodon/feed/profile/toot/Context.kt
@@ -1,6 +1,5 @@
 package com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot
 
-import com.jeanbarrossilva.orca.core.mastodon.feed.profile.toot.status.Status
 import kotlinx.serialization.Serializable
 
 @Serializable
